\name{hsMirrorLocs}
\alias{hsMirrorLocs}
\docType{data}
\title{Mirror locations of the human genome}
\description{
Mirror locations of the human genome, based on the NCBI 36 assembly of the human genome, for use with the KCsmart package.
}
\usage{hsMirrorLocs}
\format{A list containing for each chromosome the start and end position and the centromere location (if a centromere is present).}
\source{Ensembl}
\references{
http://www.ensembl.org
}
\keyword{datasets}
