\name{hsSampleData}
\alias{hsSampleData}
\docType{data}
\title{Homo Sapiens artificial cgh data set}
\description{
An artificial cgh data set, created by permuting a BAC data set consisting of 20 samples and introducing an artificial gain on 1p.
To be used with the KCsmart package.
}
\usage{hsSampleData}
\format{A data.frame containing 3268 rows and 22 columns}
\source{Artificial data set}
\references{
}
\keyword{datasets}
