\name{KCData-class}
\docType{class}
\alias{KCData-class}
\alias{[[<-,KCData-method}
\alias{[[,KCData-method}
\alias{length,KCData-method}
\alias{unlist,KCData-method}

\title{Internal class "KCData" }
\description{Internal superclass "KCData"}
\section{Objects from the Class}{
	The user is not meant to create instances of this class 
}
\section{Slots}{
	 \describe{
    \item{\code{data}:}{Internal data}
  }
}
\section{Methods}{
  \describe{
    \item{[[<-}{\code{signature(x = "KCData")}: ... }
    \item{[[}{\code{signature(x = "KCData")}: ... }
    \item{length}{\code{signature(x = "KCData")}: ... }
    \item{unlist}{\code{signature(x = "KCData")}: ... }
	 }
}
\author{Jorma de Ronde}
\note{Internal class }

\section{Warning }{This class is meant for internal use only}

\keyword{classes}
