\name{KcghData-class}
\docType{class}
\alias{KcghData-class}
\alias{initialize,KcghData-method}
\alias{sort,KcghData-method}

\title{Class "KcghData" }
\description{Internal class}

\section{Slots}{
    \describe{
    \item{\code{probeAnnotation}:}{Object of class \code{"probeAnnotation"}}
    \item{\code{data}:}{Holds aCGH data}
  }
}

\section{Methods}{
  \describe{
    \item{initialize}{\code{signature(.Object = "KcghData")}: Internal use only }
    \item{sort}{\code{signature(x = "KcghData")}: Internal use only }
	 }
}

\author{Jorma de Ronde }
\note{For internal use only }

\keyword{classes}
