\name{KcghDataSplit-class}
\docType{class}
\alias{KcghDataSplit-class}
\alias{initialize,KcghDataSplit-method}

\title{Class "KcghDataSplit"}
\description{Internal class}


\section{Slots}{
    \describe{
    \item{\code{probeAnnotation}:}{Object of class \code{"probeAnnotation"}}
    \item{\code{pos}:}{Holds aCGH data for losses}
    \item{\code{neg}:}{Holds aCGH data for gains}
  }
}

\section{Methods}{
  \describe{
    \item{initialize}{\code{signature(.Object = "KcghDataSplit")}: Internal use only }
	 }
}
\author{ Jorma de Ronde}
\note{For internal use only}

\keyword{classes}
