\name{mmMirrorLocs}
\alias{mmMirrorLocs}
\docType{data}
\title{Mirror locations of the mouse genome}
\description{
Mirror locations of the mouse genome, based on the NCBI m37 mouse assembly, for use with the KCsmart package.
}
\usage{mmMirrorLocs}
\format{A list containing for each chromosome the start and end position.}
\source{Ensembl}
\references{
http://www.ensembl.org
}
\keyword{datasets}
